requirements.txt
flask
pandas
numpy
mysql-connector-python
matplotlib
seaborn
